update kttp to remove warning and use new apis
