+++ instrumentation/khttp-0.1/src/main/java/com/splunk/opentelemetry/instrumentation/khttp/KHttpHttpClientHttpAttributesGetter.java
@@ -20,9 +20,13 @@ import static java.util.Collections.emptyList;
 import static java.util.Collections.singletonList;
 
 import io.opentelemetry.instrumentation.api.instrumenter.http.HttpClientAttributesGetter;
+
+import java.net.InetSocketAddress;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
+
+import io.opentelemetry.semconv.trace.attributes.SemanticAttributes;
 import khttp.responses.Response;
 import org.jetbrains.annotations.Nullable;
 
@@ -61,4 +65,53 @@ final class KHttpHttpClientHttpAttributesGetter
     String header = response.getHeaders().get(name);
     return header != null ? singletonList(header) : emptyList();
   }
+
+  @Override
+  public String getTransport(RequestWrapper requestWrapper, @Nullable Response response) {
+    return SemanticAttributes.NetTransportValues.IP_TCP;
+  }
+
+  @Nullable
+  @Override
+  public String getServerAddress(RequestWrapper requestWrapper) {
+    if (requestWrapper.parsedUri != null) {
+      return requestWrapper.parsedUri.getHost();
+    }
+    return null;
+  }
+
+  @Nullable
+  @Override
+  public Integer getServerPort(RequestWrapper requestWrapper) {
+    if (requestWrapper.parsedUri != null && requestWrapper.parsedUri.getPort() > 0) {
+      return requestWrapper.parsedUri.getPort();
+    }
+    return null;
+  }
+
+  @Nullable
+  @Override
+  public String getNetworkProtocolName(RequestWrapper requestWrapper, @Nullable Response response) {
+    if(requestWrapper.parsedUri.getScheme().toLowerCase().startsWith("http")) {
+      return "http";
+    }
+    return null;
+  }
+
+  @Override
+  public String getNetworkTransport(RequestWrapper requestWrapper, @Nullable Response response) {
+    return "tcp";
+  }
+
+  @Nullable
+  @Override
+  public InetSocketAddress getServerInetSocketAddress(RequestWrapper requestWrapper, @Nullable Response response) {
+    String host = getServerAddress(requestWrapper);
+    Integer port = getServerPort(requestWrapper);
+    if(host == null || port == null){
+      return null;
+    }
+    return new InetSocketAddress(host, port);
+  }
+
 }
